feat: feign servcie preparing, migration fixes

+++ src/main/resources/db/migration/V1__Initial_Version.sql
@@ -9,16 +9,16 @@ CREATE TABLE IF NOT EXISTS "user"
 
 CREATE TABLE IF NOT EXISTS account
 (
-    id              BIGSERIAL PRIMARY KEY,
+    id              INTEGER PRIMARY KEY,
     username        VARCHAR(255),
     is_main_account BOOLEAN
 );
 
 CREATE TABLE IF NOT EXISTS account_linked_account
 (
-    id                BIGSERIAL PRIMARY KEY,
-    main_account_id   BIGINT,
-    linked_account_id BIGINT,
+    id                INTEGER PRIMARY KEY,
+    main_account_id   INTEGER,
+    linked_account_id INTEGER,
     FOREIGN KEY (main_account_id) REFERENCES account (id),
     FOREIGN KEY (linked_account_id) REFERENCES account (id)
 );
@@ -31,9 +31,9 @@ CREATE TABLE IF NOT EXISTS token
 
 CREATE TABLE IF NOT EXISTS subscription
 (
-    id              BIGSERIAL PRIMARY KEY,
+    id              serial PRIMARY KEY,
     expiration_date TIMESTAMP,
     is_active       BOOLEAN,
-    token_id        BIGINT,
+    token_id        INTEGER,
     FOREIGN KEY (token_id) REFERENCES token (id)
 );
